Fixed bug when drawing rectangles from directions other than UpperLeft to down and right

+++ ShapeTesterWebLayer/ViewModel/Rectangle.cs
@@ -7,12 +7,12 @@ namespace ShapeTesterWebLayer.ViewModel
 {
     public class Rectangle
     {
-        public double X { get; set; }
+        public int X { get; set; }
 
-        public double Y { get; set; }
+        public int Y { get; set; }
 
-        public double Width { get; set; }
+        public int Width { get; set; }
 
-        public double Height { get; set; }
+        public int Height { get; set; }
     }
 }
